A little refactor
Moved the Input component into its own class

Made type and placeholder optional on my Input component

Removed the firstname, lastname, ..., repeatPassword functions
in favour of just calling my Input component directly in sign-up and
log-in area. Not as big a need to remove code duplication now that
it's only the label that's required on those Input component calls.