Huge refactor
Log interface lives in PopUpForm/index.tsx

The return type of LogIn and SignUp functions (not the components,
but the onSubmit functions) is now LoginAPIReponse and
SignupAPIResponse respectively rather than Log[].

LoginAPIResponse and SignupAPIResponse both extend from
CommonAPIResponse living in PopUpForm/index.tsx. It has a logs field
of type Log[].

APIResponses renamed to APILogs in both the LogIn and SignUp
components. APILogs state continues to be of type Log[] even as
the submit functions on the two form's submit button expects to have
a parameter of type LoginAPIResponse and SignupAPIResponse
respectively. A proxy function is provided in each of the popup kinds
to mitigate this discrepancy.