Logout implemented
A rather naive solution for now. Logout is always permitted to be
performed by the client, regardless of the credentials. This is
because for now, the process of logging out is just the client
deleting its own cookies from itself. In the future with database
stored session tokens (session management), the process of logging
out would include deleting the session token on the database which
would require proper credentials.

delete cookie into its own file