images in database
images in database are now just strings.

In the react layer, there is a base64 util for ensuring
that a string abides by the character set of a base64 encoding.
It is used to call the fallback either when no image string is in
the database for a given pizza, OR when that stores string doesn't
obey the base64 restrictions.